move models into model.rs
