+++ src/core/blueprint/operators/expr.rs
@@ -3,8 +3,8 @@ use async_graphql_value::ConstValue;
 use crate::core::blueprint::*;
 use crate::core::config;
 use crate::core::config::Field;
-use crate::core::ir::IR;
-use crate::core::ir::IR::Dynamic;
+use crate::core::ir::model::IR;
+use crate::core::ir::model::IR::Dynamic;
 use crate::core::try_fold::TryFold;
 use crate::core::valid::{Valid, ValidationError, Validator};
