+++ src/core/blueprint/operators/modify.rs
@@ -1,7 +1,7 @@
 use crate::core::blueprint::*;
 use crate::core::config;
 use crate::core::config::Field;
-use crate::core::ir::{Context, IR};
+use crate::core::ir::model::{Context, IR};
 use crate::core::try_fold::TryFold;
 use crate::core::valid::Valid;
