+++ src/core/blueprint/operators/protected.rs
@@ -1,6 +1,6 @@
 use crate::core::blueprint::FieldDefinition;
 use crate::core::config::{self, ConfigModule, Field};
-use crate::core::ir::{Context, IR};
+use crate::core::ir::model::{Context, IR};
 use crate::core::try_fold::TryFold;
 use crate::core::valid::Valid;
