+++ src/core/ir/mod.rs
@@ -1,4 +1,3 @@
-mod cache;
 mod error;
 mod eval;
 mod evaluation_context;
@@ -8,53 +7,9 @@ mod io;
 mod modify;
 mod resolver_context_like;
 
-use std::collections::HashMap;
-use std::fmt::Debug;
-pub use cache::*;
+pub mod model;
 pub use error::*;
 pub use eval::*;
 pub use evaluation_context::EvaluationContext;
 pub use graphql_operation_context::GraphQLOperationContext;
-pub use io::*;
 pub use resolver_context_like::{EmptyResolverContext, ResolverContext, ResolverContextLike};
-use strum_macros::Display;
-
-use crate::core::blueprint::DynamicValue;
-
-
-#[derive(Clone, Debug, Display)]
-pub enum IR {
-    Context(Context),
-    Dynamic(DynamicValue),
-    #[strum(to_string = "{0}")]
-    IO(IO),
-    Cache(Cache),
-    Path(Box<IR>, Vec<String>),
-    Protect(Box<IR>),
-    Map(Map),
-}
-
-#[derive(Clone, Debug)]
-pub enum Context {
-    Value,
-    Path(Vec<String>),
-    PushArgs { expr: Box<IR>, and_then: Box<IR> },
-    PushValue { expr: Box<IR>, and_then: Box<IR> },
-}
-
-impl IR {
-    pub fn and_then(self, next: Self) -> Self {
-        IR::Context(Context::PushArgs { expr: Box::new(self), and_then: Box::new(next) })
-    }
-
-    pub fn with_args(self, args: IR) -> Self {
-        IR::Context(Context::PushArgs { expr: Box::new(args), and_then: Box::new(self) })
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct Map {
-    pub input: Box<IR>,
-    // accept key return value instead of
-    pub map: HashMap<String, String>,
-}
