+++ src/core/mod.rs
@@ -44,7 +44,7 @@ use std::num::NonZeroU64;
 
 use async_graphql_value::ConstValue;
 use http::Response;
-use ir::IoId;
+use ir::model::IoId;
 pub use tailcall_macros as macros;
 
 pub trait EnvIO: Send + Sync + 'static {
