+++ tailcall-cloudflare/src/cache.rs
@@ -4,7 +4,7 @@ use std::rc::Rc;
 use anyhow::Result;
 use async_graphql_value::ConstValue;
 use serde_json::Value;
-use tailcall::core::ir::IoId;
+use tailcall::core::ir::model::IoId;
 use tailcall::core::Cache;
 use worker::kv::KvStore;
